Move to Analyser2; the background to this is:
----- message from Saleae support -------
The example you provided with ClearTabularText was the correct one.

The original API was written before we implemented protocol search support, with the idea that we would add it later.

However it was obvious later during implementation that we couldn't use ClearTabularText and AddResultString from another thread at the same time protocol result bubbles were being generated. In the process we broke the binary API. (which was mainly broken by the modifications to the Channel type)

If it's not working, the problem might be that your analyzer does not inherit from Analyzer2. Please review this in the I2C source:
https://github.com/saleae/i2c-analyzer/blob/master/src/I2cAnalyzer.h#L9

Analyzer2 was created when we needed to add a setup function to the interface, allowing setup to be done from the main thread instead of the worker thread.
Implementation for I2C is here:
https://github.com/saleae/i2c-analyzer/blob/master/src/I2cAnalyzer.cpp#L18

Basically the results setup needs to be moved from the WorkerThread to the new overload.
Mark Garrison